:truck: Move namespace of IDataStore to CAFU.Core.Data.DataStore
