+++ Assets/Scripts/CAFU/Core/Data/DataStore/DataStore.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 // ReSharper disable UnusedMember.Global
 
-namespace CAFU.Core.Data {
+namespace CAFU.Core.Data.DataStore {
 
     public interface IDataStore {
     }
